# postman/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/produtos/{id}":{"get":{"tags":["produto-controller"],"operationId":"buscarPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProdutoResponse"}}}}}},"put":{"tags":["produto-controller"],"operationId":"atualizar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProdutoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProdutoResponse"}}}}}},"delete":{"tags":["produto-controller"],"operationId":"excluir","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/produto-materiais/{id}":{"put":{"tags":["produto-material-controller"],"operationId":"atualizar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProdutoMaterialUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProdutoMaterialResponse"}}}}}},"delete":{"tags":["produto-material-controller"],"operationId":"excluir_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/pessoas/{id}":{"get":{"tags":["pessoa-controller"],"operationId":"buscarPorId_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PessoaResponse"}}}}}},"put":{"tags":["pessoa-controller"],"operationId":"atualizar_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PessoaUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PessoaResponse"}}}}}},"delete":{"tags":["pessoa-controller"],"operationId":"excluir_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/ordens-producao/{id}":{"get":{"tags":["ordem-producao-controller"],"operationId":"buscarPorId_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemProducaoResponse"}}}}}},"put":{"tags":["ordem-producao-controller"],"operationId":"atualizar_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrdemProducaoUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemProducaoResponse"}}}}}},"delete":{"tags":["ordem-producao-controller"],"operationId":"excluir_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/op-materiais/{id}":{"put":{"tags":["op-material-controller"],"operationId":"atualizar_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OpMaterialUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OpMaterialResponse"}}}}}},"delete":{"tags":["op-material-controller"],"operationId":"excluir_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/movimentacoes-estoque/{id}":{"get":{"tags":["movimentacao-estoque-controller"],"operationId":"buscarPorId_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovimentacaoEstoqueResponse"}}}}}},"put":{"tags":["movimentacao-estoque-controller"],"operationId":"atualizar_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovimentacaoEstoqueUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovimentacaoEstoqueResponse"}}}}}},"delete":{"tags":["movimentacao-estoque-controller"],"operationId":"excluir_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/materiais/{id}":{"get":{"tags":["material-controller"],"operationId":"buscarPorId_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MaterialResponse"}}}}}},"put":{"tags":["material-controller"],"operationId":"atualizar_6","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaterialUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MaterialResponse"}}}}}},"delete":{"tags":["material-controller"],"operationId":"excluir_6","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/locais/{id}":{"get":{"tags":["local-controller"],"operationId":"buscarPorId_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocalResponse"}}}}}},"put":{"tags":["local-controller"],"operationId":"atualizar_7","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocalUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocalResponse"}}}}}},"delete":{"tags":["local-controller"],"operationId":"excluir_7","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/apontamentos-producao/{id}":{"get":{"tags":["apontamento-producao-controller"],"operationId":"buscarPorId_6","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApontamentoProducaoResponse"}}}}}},"put":{"tags":["apontamento-producao-controller"],"operationId":"atualizar_8","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApontamentoProducaoUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApontamentoProducaoResponse"}}}}}},"delete":{"tags":["apontamento-producao-controller"],"operationId":"excluir_8","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/tipos-pessoa":{"get":{"tags":["tipo-pessoa-controller"],"operationId":"listar","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TipoPessoaResponse"}}}}}}},"post":{"tags":["tipo-pessoa-controller"],"operationId":"criar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TipoPessoaRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TipoPessoaResponse"}}}}}}},"/produtos":{"get":{"tags":["produto-controller"],"operationId":"listar_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProdutoResponse"}}}}}}},"post":{"tags":["produto-controller"],"operationId":"salvar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProdutoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProdutoResponse"}}}}}}},"/produto-materiais":{"post":{"tags":["produto-material-controller"],"operationId":"criar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProdutoMaterialRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProdutoMaterialResponse"}}}}}}},"/pessoas":{"get":{"tags":["pessoa-controller"],"operationId":"listar_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PessoaResponse"}}}}}}},"post":{"tags":["pessoa-controller"],"operationId":"criar_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PessoaRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PessoaResponse"}}}}}}},"/pessoas/{id}/tipos":{"post":{"tags":["pessoa-tipo-controller"],"operationId":"adicionarTipo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PessoaTipoRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/ordens-producao":{"get":{"tags":["ordem-producao-controller"],"operationId":"listar_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrdemProducaoResponse"}}}}}}},"post":{"tags":["ordem-producao-controller"],"operationId":"criar_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrdemProducaoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemProducaoResponse"}}}}}}},"/ordens-producao/{id}/reservar":{"post":{"tags":["ordem-producao-controller"],"operationId":"reservar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemProducaoResponse"}}}}}}},"/ordens-producao/{id}/reabrir":{"post":{"tags":["ordem-producao-controller"],"operationId":"reabrir","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemProducaoResponse"}}}}}}},"/ordens-producao/{id}/iniciar":{"post":{"tags":["ordem-producao-controller"],"operationId":"iniciar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemProducaoResponse"}}}}}}},"/ordens-producao/{id}/falhar":{"post":{"tags":["ordem-producao-controller"],"operationId":"falhar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemProducaoResponse"}}}}}}},"/ordens-producao/{id}/concluir":{"post":{"tags":["ordem-producao-controller"],"operationId":"concluir","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemProducaoResponse"}}}}}}},"/ordens-producao/{id}/cancelar":{"post":{"tags":["ordem-producao-controller"],"operationId":"cancelar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemProducaoResponse"}}}}}}},"/op-materiais":{"post":{"tags":["op-material-controller"],"operationId":"criar_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OpMaterialRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OpMaterialResponse"}}}}}}},"/movimentacoes-estoque":{"get":{"tags":["movimentacao-estoque-controller"],"operationId":"listar_4","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovimentacaoEstoqueResponse"}}}}}}},"post":{"tags":["movimentacao-estoque-controller"],"operationId":"criar_5","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovimentacaoEstoqueRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovimentacaoEstoqueResponse"}}}}}}},"/materiais":{"get":{"tags":["material-controller"],"operationId":"listar_5","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MaterialResponse"}}}}}}},"post":{"tags":["material-controller"],"operationId":"criar_6","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaterialRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MaterialResponse"}}}}}}},"/locais":{"get":{"tags":["local-controller"],"operationId":"listar_6","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LocalResponse"}}}}}}},"post":{"tags":["local-controller"],"operationId":"salvar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocalRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocalResponse"}}}}}}},"/categorias":{"get":{"tags":["categoria-controller"],"operationId":"listar_7","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Categoria"}}}}}}},"post":{"tags":["categoria-controller"],"operationId":"salvar_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Categoria"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Categoria"}}}}}}},"/apontamentos-producao":{"post":{"tags":["apontamento-producao-controller"],"operationId":"criar_7","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApontamentoProducaoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApontamentoProducaoResponse"}}}}}}},"/produtos/inativos":{"get":{"tags":["produto-controller"],"operationId":"listarInativos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProdutoResponse"}}}}}}}},"/produto-materiais/produto/{produtoId}":{"get":{"tags":["produto-material-controller"],"operationId":"listarPorProduto","parameters":[{"name":"produtoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProdutoMaterialResponse"}}}}}}}},"/op-materiais/ordem/{ordemProducaoId}":{"get":{"tags":["op-material-controller"],"operationId":"listarPorOp","parameters":[{"name":"ordemProducaoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OpMaterialResponse"}}}}}}}},"/estoque":{"get":{"tags":["estoque-controller"],"operationId":"listarSaldo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EstoqueResponse"}}}}}}}},"/estoque/disponivel/{materialId}":{"get":{"tags":["estoque-controller"],"operationId":"consultarDisponivel","parameters":[{"name":"materialId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"number"}}}}}}},"/apontamentos-producao/ordem/{ordemProducaoId}":{"get":{"tags":["apontamento-producao-controller"],"operationId":"listarPorOp_1","parameters":[{"name":"ordemProducaoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApontamentoProducaoResponse"}}}}}}}}},"components":{"schemas":{"ProdutoRequest":{"type":"object","properties":{"nome":{"type":"string","minLength":1},"descricao":{"type":"string"},"categoriaId":{"type":"integer","format":"int64"}},"required":["nome"]},"ProdutoResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"codigo":{"type":"string"},"nome":{"type":"string"},"descricao":{"type":"string"},"categoria":{"type":"string"},"ativo":{"type":"boolean"}}},"ProdutoMaterialUpdateRequest":{"type":"object","properties":{"quantidade":{"type":"number","minimum":0.001},"ativo":{"type":"boolean"}}},"ProdutoMaterialResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"produtoId":{"type":"integer","format":"int64"},"produto":{"type":"string"},"materialId":{"type":"integer","format":"int64"},"material":{"type":"string"},"quantidade":{"type":"number"},"ativo":{"type":"boolean"},"dataCriacao":{"type":"string","format":"date-time"}}},"PessoaUpdateRequest":{"type":"object","properties":{"nome":{"type":"string"},"tipoDocumento":{"type":"string"},"documento":{"type":"string"},"telefone":{"type":"string"},"email":{"type":"string"},"observacao":{"type":"string"},"ativo":{"type":"boolean"}}},"PessoaResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"tipoDocumento":{"type":"string"},"documento":{"type":"string"},"telefone":{"type":"string"},"email":{"type":"string"},"observacao":{"type":"string"},"ativo":{"type":"boolean"},"dataCriacao":{"type":"string","format":"date-time"},"tipos":{"type":"array","items":{"type":"string"}}}},"OrdemProducaoUpdateRequest":{"type":"object","properties":{"quantidadePlanejada":{"type":"number"},"quantidadeProduzida":{"type":"number"},"localDestinoId":{"type":"integer","format":"int64"},"responsavelId":{"type":"integer","format":"int64"},"origem":{"type":"string"},"observacao":{"type":"string"},"ativo":{"type":"boolean"}}},"OrdemProducaoResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"numero":{"type":"string"},"produtoId":{"type":"integer","format":"int64"},"produto":{"type":"string"},"quantidadePlanejada":{"type":"number"},"quantidadeProduzida":{"type":"number"},"localDestinoId":{"type":"integer","format":"int64"},"localDestino":{"type":"string"},"responsavelId":{"type":"integer","format":"int64"},"responsavel":{"type":"string"},"status":{"type":"string"},"origem":{"type":"string"},"observacao":{"type":"string"},"ativo":{"type":"boolean"},"dataAbertura":{"type":"string","format":"date-time"},"dataInicio":{"type":"string","format":"date-time"},"dataFim":{"type":"string","format":"date-time"}}},"OpMaterialUpdateRequest":{"type":"object","properties":{"quantidadePlanejada":{"type":"number","minimum":0.001},"quantidadeConsumida":{"type":"number","minimum":0.000},"ativo":{"type":"boolean"}}},"OpMaterialResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"ordemProducaoId":{"type":"integer","format":"int64"},"ordemProducao":{"type":"string"},"materialId":{"type":"integer","format":"int64"},"material":{"type":"string"},"quantidadePlanejada":{"type":"number"},"quantidadeConsumida":{"type":"number"},"ativo":{"type":"boolean"},"dataCriacao":{"type":"string","format":"date-time"}}},"MovimentacaoEstoqueUpdateRequest":{"type":"object","properties":{"quantidade":{"type":"number","minimum":0.001},"observacao":{"type":"string"},"ativo":{"type":"boolean"}}},"MovimentacaoEstoqueResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"produtoId":{"type":"integer","format":"int64"},"produto":{"type":"string"},"materialId":{"type":"integer","format":"int64"},"material":{"type":"string"},"localId":{"type":"integer","format":"int64"},"local":{"type":"string"},"tipo":{"type":"string"},"movimento":{"type":"string"},"quantidade":{"type":"number"},"origem":{"type":"string"},"referenciaId":{"type":"integer","format":"int64"},"observacao":{"type":"string"},"responsavelId":{"type":"integer","format":"int64"},"responsavel":{"type":"string"},"dataMovimentacao":{"type":"string","format":"date-time"},"ativo":{"type":"boolean"}}},"MaterialUpdateRequest":{"type":"object","properties":{"nome":{"type":"string"},"descricao":{"type":"string"},"unidadeMedida":{"type":"string"},"estoqueMinimo":{"type":"number"},"ativo":{"type":"boolean"}}},"MaterialResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"codigo":{"type":"string"},"nome":{"type":"string"},"descricao":{"type":"string"},"unidadeMedida":{"type":"string"},"estoqueMinimo":{"type":"number"},"ativo":{"type":"boolean"},"dataCriacao":{"type":"string","format":"date-time"}}},"LocalUpdateRequest":{"type":"object","properties":{"nome":{"type":"string","minLength":1},"ativo":{"type":"boolean"}},"required":["nome"]},"LocalResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"ativo":{"type":"boolean"},"dataCriacao":{"type":"string","format":"date-time"}}},"ApontamentoProducaoUpdateRequest":{"type":"object","properties":{"quantidadeProduzida":{"type":"number","minimum":0.000},"quantidadePerda":{"type":"number","minimum":0.000},"observacao":{"type":"string"},"ativo":{"type":"boolean"}}},"ApontamentoProducaoResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"ordemProducaoId":{"type":"integer","format":"int64"},"ordemProducao":{"type":"string"},"quantidadeProduzida":{"type":"number"},"quantidadePerda":{"type":"number"},"responsavelId":{"type":"integer","format":"int64"},"responsavel":{"type":"string"},"observacao":{"type":"string"},"dataApontamento":{"type":"string","format":"date-time"},"ativo":{"type":"boolean"}}},"TipoPessoaRequest":{"type":"object","properties":{"nome":{"type":"string","minLength":1},"descricao":{"type":"string"}},"required":["nome"]},"TipoPessoaResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"descricao":{"type":"string"},"ativo":{"type":"boolean"},"dataCriacao":{"type":"string","format":"date-time"}}},"ProdutoMaterialRequest":{"type":"object","properties":{"produtoId":{"type":"integer","format":"int64"},"materialId":{"type":"integer","format":"int64"},"quantidade":{"type":"number","minimum":0.001}},"required":["materialId","produtoId","quantidade"]},"PessoaRequest":{"type":"object","properties":{"nome":{"type":"string","minLength":1},"tipoDocumento":{"type":"string"},"documento":{"type":"string"},"telefone":{"type":"string"},"email":{"type":"string"},"observacao":{"type":"string"}},"required":["nome"]},"PessoaTipoRequest":{"type":"object","properties":{"tipoPessoaId":{"type":"integer","format":"int64"}},"required":["tipoPessoaId"]},"OrdemProducaoRequest":{"type":"object","properties":{"numero":{"type":"string","minLength":1},"produtoId":{"type":"integer","format":"int64"},"quantidadePlanejada":{"type":"number","minimum":0.001},"localDestinoId":{"type":"integer","format":"int64"},"responsavelId":{"type":"integer","format":"int64"},"origem":{"type":"string","minLength":1},"observacao":{"type":"string"}},"required":["localDestinoId","numero","origem","produtoId","quantidadePlanejada"]},"OpMaterialRequest":{"type":"object","properties":{"ordemProducaoId":{"type":"integer","format":"int64"},"materialId":{"type":"integer","format":"int64"},"quantidadePlanejada":{"type":"number","minimum":0.001}},"required":["materialId","ordemProducaoId","quantidadePlanejada"]},"MovimentacaoEstoqueRequest":{"type":"object","properties":{"produtoId":{"type":"integer","format":"int64"},"materialId":{"type":"integer","format":"int64"},"localId":{"type":"integer","format":"int64"},"tipo":{"type":"string"},"movimento":{"type":"string"},"quantidade":{"type":"number","minimum":0.001},"origem":{"type":"string"},"referenciaId":{"type":"integer","format":"int64"},"observacao":{"type":"string"},"responsavelId":{"type":"integer","format":"int64"}},"required":["localId","origem","quantidade","tipo"]},"MaterialRequest":{"type":"object","properties":{"codigo":{"type":"string","minLength":1},"nome":{"type":"string","minLength":1},"descricao":{"type":"string"},"unidadeMedida":{"type":"string","minLength":1},"estoqueMinimo":{"type":"number"}},"required":["codigo","estoqueMinimo","nome","unidadeMedida"]},"LocalRequest":{"type":"object","properties":{"nome":{"type":"string","minLength":1}},"required":["nome"]},"Categoria":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"ativo":{"type":"boolean"},"dataCriacao":{"type":"string","format":"date-time"}}},"ApontamentoProducaoRequest":{"type":"object","properties":{"ordemProducaoId":{"type":"integer","format":"int64"},"quantidadeProduzida":{"type":"number","minimum":0.000},"quantidadePerda":{"type":"number","minimum":0.000},"responsavelId":{"type":"integer","format":"int64"},"observacao":{"type":"string"}},"required":["ordemProducaoId","quantidadePerda","quantidadeProduzida"]},"EstoqueResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"codigo":{"type":"string"},"nome":{"type":"string"},"tipo":{"type":"string"},"localId":{"type":"integer","format":"int64"},"local":{"type":"string"},"saldo":{"type":"number"}}}}}}